+++ packages/tide_kit/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Next
+
+- Added example 2 which is a Spotify search app.
+- Added the TidePanelTitleBar, TideSearchPanel, and TideSearchTextField widgets.
+
 ## 1.0.1
 
 * Minor pedantic changes.
@@ -5,4 +10,3 @@
 ## 1.0.0
 
 * The initial release of Tide Kit. Tide Kit is a toolkit for building IDE and studio like apps on all platforms using Flutter.
-